1.09  + add skt default settings for all http server bindings  + add new option: fKeepAlive  + add new option: keepAliveIDLEtime  + add new option: keepAliveProbesInterval  + add new option: keepAliveNumberOfProbes  * http server lib upgraded to 1.6  * fix: http server: 'Content-Size' -> 'Content-Length'  * fix: linux: apply system limits before create thread pool  * fix: FreeBSD: thread pool may lost messages  * fAcceptFilter http server param does not work  * mmalloc_fd cant allocate >50% mem on FreeBSD without increace vm.max_wired  - remove option: fAcceptFilter for http server - allways ON

+++ src/include/core_dns_resolv.h
@@ -1,5 +1,5 @@
 /*-
- * Copyright (c) 2011 - 2014 Rozhuk Ivan <rozhuk.im@gmail.com>
+ * Copyright (c) 2011 - 2016 Rozhuk Ivan <rozhuk.im@gmail.com>
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -39,8 +39,8 @@
 typedef struct dns_rslvr_s	*dns_rslvr_p; /* thread pool */
 typedef struct dns_rslvr_task_s	*dns_rslvr_task_p; /* thread pool */
 
-#define DNS_R_F_IPV4	(1 << 0) /* IPv4 */
-#define DNS_R_F_IPV6	(1 << 1) /* IPv6 */
+#define DNS_R_F_IPV4	(((uint16_t)1) << 0) /* IPv4 */
+#define DNS_R_F_IPV6	(((uint16_t)1) << 1) /* IPv6 */
 #define DNS_R_F_IP_ALL	(DNS_R_F_IPV4 | DNS_R_F_IPV6)
